srm: Change job ID type from Long to long
A recent patch replaced the canonicalization code such that it no
longer relies on a reference to a particular instance of Long. Thus
we can now use long rather than Long to represent required job IDs.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5946/

+++ modules/srm/src/main/java/org/dcache/srm/request/Request.java
@@ -124,7 +124,7 @@ public abstract class Request extends Job {
      */
 
     protected Request(
-    Long id,
+    long id,
     Long nextJobId,
     long creationTime,
     long lifetime,
@@ -243,7 +243,7 @@ public abstract class Request extends Job {
      * request id
      */
     public int getRequestNum() {
-        return (int) (getId().longValue());
+        return (int) getId();
     }
